+++ nflow-engine/src/main/java/io/nflow/engine/config/db/H2DatabaseConfiguration.java
@@ -82,14 +82,6 @@ public class H2DatabaseConfiguration extends DatabaseConfiguration {
     }
 
     /**
-     * Returns false as H2 does not support update returning clause.
-     */
-    @Override
-    public boolean hasUpdateReturning() {
-      return false;
-    }
-
-    /**
      * Returns false as H2 does not support updateable CTEs.
      */
     @Override