+++ nflow-engine/src/main/java/io/nflow/engine/internal/storage/db/MySQLVariants.java
@@ -18,14 +18,6 @@ public class MySQLVariants implements SQLVariants {
   }
 
   /**
-   * Returns false as MySQL does not support update returning clause.
-   */
-  @Override
-  public boolean hasUpdateReturning() {
-    return false;
-  }
-
-  /**
    * Returns false as MySQL does not support updateable CTEs.
    */
   @Override